TWI Refactor `print()` away from a Visitor pattern
`sealed interface Expr` allows us to use an exhaustive `when` statement
in `print()` instead of an explicit `visitor.accept()`. The
implementations are now simple data aggregators.

`print()` itself need not be a method of `AstPrinter()`, and this class
can be removed. Rename the now free function to `printAst()` to avoid
confusing it with `Kotlin.stdlib.print()`.

This makes `GenerateAST.py` somewhat redundant, since it is barely any
work to manually add the new expressions, or so I think. I leave the
script since as insurance if things get more complicated later on.