TWI "Evaluating Expressions"
The book defines an `Interpreter` class, so that it can be initialised
only once, and it claims it will become beneficial in later chapters.
I am intentionally making `interpret()` a free function since we do not
yet need to store any state. I will update it if this becomes an issue.

Remember that we are using an alternative to the visitor pattern that
uses a `sealed interface`. Instead of inheriting from a Visitor class,
we can have it so that `evaluate()` simply switches between all the
appropriate `data classes`.

Skip unit tests simply because I am lazy, and they are not "mandatory"
material from the book. I did do some basic manual tests, and we can
interpret tokens as intended.