TWI "Statements and State: Statements"
Both `expressionStatement()` and `printStatement()` must first evaluate
the passed expression before consuming the end-of-statement character
(';'); so, we must store both `expr`, and `value` even if they are used
only once, and could otherwise be inlined.

Pass `base_class` when defining a type, since not all data classes
depend on `Expr`, as it was previously hardcoded.