TWI "Statements and State: Global Variables", and ": Environment"
We now see why "Crafting Interpreters" defines an `Interpreter` object,
with state: (at least) to store an `Environment`. I want to follow the
sunken costs fallacy, and define this instead on a global level. We thus
have a single `Environment` for the entire program; I am curious to see
where this becomes a problem.