TWI "Statements and State: Scope"
Seems to still work with global `environment`, and not need an enclosing
object. Let us keep pushing this global `environment` "fallacy"!

Chaining `.also {}` seems to work as intended, although I am unsure as
to how it is actually doing so: we first also change the global
`environment`; then also execute the statements; and then finally also
change the global once again to the original referenced object? Since
the referenced object is global, I assume we make a copy of
`environment` when we chain it with these scoped functions. This is not
something I would like to use in actual production code, until I am sure
of how this works in detail!